+++ resources/course.py
@@ -1,5 +1,6 @@
 from flask_restful import Resource
 from flask import request
+from flask_jwt_extended import get_jwt_identity, jwt_required
 
 from models.course import CourseModel
 from schemas.course import CourseSchema
@@ -17,6 +18,7 @@ class Course(Resource):
         return {'message': 'Course not found'}, 404
 
     @classmethod
+    @jwt_required()
     def post(cls, slug: str) -> any:
         if CourseModel.find_by_name(slug):
             return {'message': 'Course with this slug already exists'}, 400
@@ -26,6 +28,8 @@ class Course(Resource):
 
         course = course_schema.load(course_json)
 
+        course.user_id = get_jwt_identity()
+
         try:
             course.save()
         except Exception as e:
@@ -34,8 +38,12 @@ class Course(Resource):
         return course_schema.dump(course), 201
 
     @classmethod
+    @jwt_required()
     def delete(cls, slug: str) -> any:
         if course := CourseModel.find_by_slug(slug):
+            if course.user_id != get_jwt_identity():
+                return {'message': 'Access denied'}, 403
+
             try:
                 course.delete()
             except Exception as e:
@@ -46,10 +54,14 @@ class Course(Resource):
         return {'message': 'Course not found'}, 404
 
     @classmethod
+    @jwt_required()
     def put(cls, slug: str) -> any:
         course_json = request.get_json()
 
         if course := CourseModel.find_by_slug(slug):
+            if course.user_id != get_jwt_identity():
+                return {'message': 'Access denied'}, 403
+
             course.description = course_json['description']
             course.video_url = course_json['video_url']
